+++ src/classes/Box.h
@@ -3,13 +3,13 @@
 class Box
 {
     public:
-        int x0;
-        int y0;
-        int xlen;
-        int ylen;
+        float x0;
+        float y0;
+        float xlen;
+        float ylen;
         int ID;
 
-        Box(int _x0, int _y0, int _xlen, int _ylen, int _ID);
+        Box(float _x0, float _y0, float _xlen, float _ylen, int _ID);
 
         void printBoxInfo();
